lab8/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.15
pydantic==1.10.7
redis==4.5.5
fakeredis==2.10.1
pytest==7.4.0
pytest-asyncio==0.23.2
httpx==0.25.0